reuse witness_gen, prove, and verify
